+++ config/easy/CouplesPictureFrame/code/CGCode.js
@@ -1,0 +1,7 @@
+const wife = readline();
+const husband = readline();
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('The Picture Frame');