+++ config/easy/CouplesPictureFrame/code/CGCode.php
@@ -1,0 +1,8 @@
+<?php
+$wife = stream_get_line(STDIN, 25 + 1, "\n");
+$husband = stream_get_line(STDIN, 25 + 1, "\n");
+
+// Write an answer using echo(). DON'T FORGET THE TRAILING \n
+// To debug: error_log(var_export($var, true)); (equivalent to var_dump)
+
+echo("The Picture Frame\n");